change main_efc_thd to main_thd because efc is not the unique correction method now. add comments
